update kurento; add use_ssh_args to all synchronize tasks; add playbooks to setup properly ufw with docker

+++ roles/bigbluebutton/files/utils/update-for-projudi.rb
@@ -8,17 +8,25 @@ require 'nokogiri'
 require 'logger'
 
 # example:
-# ruby update-for-projudi.rb -x "/recording/meta/bbb-context" -v "2º Juizado Especial Cível"
+# ruby update-for-projudi.rb --xpath "/recording/meta/bbb-context" --value "2º Juizado Especial Cível"
+# ruby update-for-projudi.rb --xpath "//meta/meetingId" --value "fc862dd423e83d64da82fbbf448caaa6c3e55097" --meeting-id "0018323-51.2016.8.23.0010" --dry-run
+# cat /var/log/bigbluebutton/update-for-projudi.log | grep -Ev "[[:digit:]]{7}-[[:digit:]]{2}\.[[:digit:]]{4}\.[[:digit:]]+\.[[:digit:]]+\.[[:digit:]]{3,4}"
 
 opts = Trollop::options do
   opt :xpath, "xpath to select recordings", :type => String, :required => true
   opt :value, "value for the xpath to match", :type => String, :required => true
+  opt :meeting_id, "new meetingId for the recordings", :type => String
+  opt :dry_run, "do not execute anything, just search", :type => :flag, :default => false
 end
 
-logger = Logger.new("/var/log/bigbluebutton/update-for-projudi.log")
+logger = if opts[:dry_run]
+        Logger.new(STDOUT)
+    else
+        Logger.new("/var/log/bigbluebutton/update-for-projudi.log")
+    end
 logger.level = Logger::INFO
 
-logger.info "Migrating \"#{opts[:value]}\""
+logger.info "Migrating query \"#{opts[:xpath]}\" value \"#{opts[:value]}\""
 
 files = `find /var/bigbluebutton/published/ /var/bigbluebutton/unpublished/ -name metadata.xml`.split("\n")
 files.each do |filename|
@@ -29,20 +37,26 @@ files.each do |filename|
         xml_node = doc.at_xpath("/recording/id")
         record_id = xml_node.content
 
-        xml_node = doc.at_xpath("/recording/meta/meetingName")
-        if xml_node.nil?
-            logger.error "Couldn't find /recording/meta/meetingName for #{record_id}, abort"
-            next
+        process_number = nil
+        if opts[:meeting_id].nil?
+            xml_node = doc.at_xpath("/recording/meta/meetingName")
+            if xml_node.nil?
+                logger.error "Couldn't find /recording/meta/meetingName for #{record_id}, abort"
+                next
+            end
+            process_number = xml_node.content
+        else
+            process_number = opts[:meeting_id]
         end
-        process_number = xml_node.content
-        logger.info "#{process_number} => #{filename}"
 
         xml_node = doc.at_xpath("/recording/meta/mconflb-institution-name")
         if xml_node.nil?
             logger.warn "Couldn't find /recording/meta/mconflb-institution-name for #{record_id}"
         else
+            next if xml_node.content == "Projudi"
             xml_node.content = "Projudi"
         end
+        logger.info "#{process_number} => #{filename}"
 
         xml_node = doc.at_xpath("/recording/meta/mconflb-institution")
         if xml_node.nil?
@@ -65,11 +79,13 @@ files.each do |filename|
             xml_node.content = process_number
         end
 
-        datetime = DateTime.parse(File.mtime(filename).to_s).strftime("%Y-%m-%d.%H%M%S")
-        FileUtils.cp filename, "#{filename}.#{datetime}"
+        if ! opts[:dry_run]
+            datetime = DateTime.parse(File.mtime(filename).to_s).strftime("%Y-%m-%d.%H%M%S")
+            FileUtils.cp filename, "#{filename}.#{datetime}"
 
-        xml_file = File.new(filename, "w")
-        xml_file.write(doc.to_xml(:indent => 2))
-        xml_file.close
+            xml_file = File.new(filename, "w")
+            xml_file.write(doc.to_xml(:indent => 2))
+            xml_file.close
+        end
     end
 end
